Fix build break on UWP (#2786)

+++ src/Native/Runtime/windows/PalRedhawkMinWin.cpp
@@ -439,7 +439,7 @@ REDHAWK_PALEXPORT void PalPrintFatalError(const char* message)
     // Write the message using lowest-level OS API available. This is used to print the stack overflow
     // message, so there is not much that can be done here.
     DWORD dwBytesWritten;
-    WriteFile(GetStdHandle(STD_ERROR_HANDLE), message, strlen(message), &dwBytesWritten, NULL);
+    WriteFile(GetStdHandle(STD_ERROR_HANDLE), message, (DWORD)strlen(message), &dwBytesWritten, NULL);
 }
 
 //
